Update structure from TOneV2_Dev
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C57386

+++ TOneV2/Code/TOneV2/SQL.TOneWhS_BE/TOneWhS_BE/Stored Procedures/sp_SupplierCode_GetByZoneIds.sql
@@ -10,9 +10,9 @@ AS
 BEGIN
 	SET NOCOUNT ON;
 
-	DECLARE @ZonesIDsTable TABLE (ZoneID INT)
+	DECLARE @ZonesIDsTable TABLE (ZoneID bigint)
 	INSERT INTO @ZonesIDsTable (ZoneID)
-	SELECT CONVERT(INT, ParsedString) FROM [TOneWhS_BE].[ParseStringList](@ZonesIDs)
+	SELECT CONVERT(bigint, ParsedString) FROM [TOneWhS_BE].[ParseStringList](@ZonesIDs)
 		
 	SELECT  [ID],[Code],[ZoneID],[BED],[EED],[CodeGroupID],[SourceID]
 	FROM	[TOneWhS_BE].[SupplierCode] sc WITH(NOLOCK)
